+++ asc-server/src/main.rs
@@ -6,11 +6,25 @@ use eyre::Result;
 use libasc::{repository::Repository, sync::{server::handle_server, stream::StdinStdout}};
 use tokio::sync::Mutex;
 
+macro_rules! error {
+    ($($t:tt)*) => {{
+        eprintln!($($t)*);
+
+        return Ok(());
+    }};
+}
+
 async fn run() -> Result<()> {
     stable_eyre::install()?;
 
+    let args: Vec<_> = std::env::args().skip(1).collect();
+
+    let Some(repo_path) = args.first() else {
+        error!("Error: repository path was not specified.");
+    };
+
     let repo = Arc::new(Mutex::new(
-        Repository::load_from("/tmp/test-remote-repo")?
+        Repository::load_from(repo_path)?
     ));
 
     let mut stream = StdinStdout::new();
@@ -18,28 +32,36 @@ async fn run() -> Result<()> {
     handle_server(&mut stream, repo).await
 }
 
-#[tokio::main]
-async fn main() {
-    if let Err(e) = run().await {
-        let mut now = Utc::now().to_string();
+fn save_error(error: &eyre::Report) {
+    let mut now = Utc::now().to_string();
 
-        if let Some(i) = now.find('.') {
-            let _ = now.split_off(i);
-        }
-
-        let name = format!("asc-server-{}", now);
-
-        let Some(dirs) = BaseDirs::new() else {
-            eprintln!("Failed to identify user directories through `directories` crate.");
-
-            return;
-        };
-
-        let log_path = dirs.config_dir().join(name);
-
-        let _ = fs::write(
-            log_path,
-            format!("{e:?}")
-        );
+    if let Some(i) = now.find('.') {
+        let _ = now.split_off(i);
     }
+
+    let name = format!("asc-server-{}", now);
+
+    let Some(dirs) = BaseDirs::new() else {
+        eprintln!("Failed to identify user directories through `directories` crate.");
+
+        return;
+    };
+
+    let log_path = dirs.cache_dir().join(name);
+
+    let _ = fs::write(
+        log_path,
+        format!("{error:?}")
+    );
+}
+
+#[tokio::main(flavor = "current_thread")]
+async fn main() -> Result<()> {
+    if let Err(e) = run().await {
+        save_error(&e);
+
+        error!("{e}");
+    }
+
+    Ok(())
 }