+++ libasc/src/sync/mod.rs
@@ -3,5 +3,6 @@ pub mod clone;
 pub mod server;
 pub mod pull; // TODO: make sure that tags also send their objects if necessary
 pub mod push;
+pub mod remote;
 pub mod stream;
 pub mod utils;